initial version of API request handler
